Pushing fixes + changed wall crush return value
Pushing against a wall would cause the player's animation to alternate between their "pushing" and "walking" ones. This was because the player's "wall_push" variable would be marked as false before their ground movement step, but would be marked as true upon colliding with a wall, and would thus alternate between the two values every step. This has been fixed by resetting their "wall_push" variable upon transitioning to their running state.

There was also an issue in which the player could push a wall without pressing the appropriate directional key; an easy way of doing so would be to launch a spindash and roll against a wall. This was because the conditions for pushing against a wall do not explicitly check if the horizontal input axis value ("input_x_direction" variable) is the same as the direction of the colliding wall; this has been fixed by adding such a check.

A minor change was also made to the wall crushing procedure: since this would kill the player, thereby changing their state, the "player_movement_*()" functions now return false in this event instead of true.