Refactored player object collision script
Originally, there was a possible memory leak caused by the DS list storing the IDs of culled objects the player was in collision with not being destroyed before the script's return values; the script has been rewritten to fix this.

A further optimization was also made by only looping through the number of colliding instances if there were any to begin with.