Rewrote save menu display
save_preview returns a ds_map that takes bits of information from an existing save file. This map is then used by the menu rather than pulling from the saves in their entirety.

Still need to work on save_default to give appropriate names, and the menu doesn't/cannot currently be refreshed to account for the newly created save.