Menu can now load saves
Modified the Accept check in the save menu to check the preview map.

save?_exists is also stored differently now, storing a number that details if the save exists and if there's a mismatch. Menu will display if there's a game/version mismatch.

Some changes to the preexisting save scripts to have local variables for the arguments.