Sensor extension and layer sensor collision fixes
The angle calculation method for solids with shapes marked as undefined has been rewritten in the same fashion as SSE V1.1, but with a fix to prevent the sensors extending beyond the given solid's bounds. The sensor are also extended up to a maximum of 4 times the player's vertical radius.
The old code for sensor extension has been commented out, if the possibility for returning to it arises in the future.

There was also an issue regarding the player's interaction with the layer sensor object; when landing on the steepest edges of the quarter-pipes that make up a big loop, the player's rotated mask could trigger the reaction of the layer sensor, even though their masks seemingly don't overlap. A fix has been included from SonicForGMS by checking if the player's center point is within the layer sensor instance before changing the player's collision layer.

Another change was made where the small quarter-pipe beside the big loop in the basic test stage was moved to the tile next to it of the same size.