Minor updates to player effect rendering
The creation of the player's effects, when applicable, has been slightly refactored, and the usage of "instance_exists" when rendering these effects has been removed due to their usage of "with," which internally verifies the existence of the provided instance.

Additionally, the spindash smoke effect's Draw event has been slightly refactored to fully utilize the "player_rotation" local variable.