Collision edits
Rewrote player_react to return true or false. If the object is a child of par_terrrain, returns false should the object be destroyed or it modified the player's speed. Otherwise returns true if the state has changed, and defaults to returning false.

Expanded constants for colliding with terrain to be more specific to the wall, floor, and ceiling.