Fixed "in_view" extension for right & bottom sides
The calculations for the right and bottom sides of the view resulted in a net extension of 0 due to a lack of consideration for the "padding" argument shifting the left and top sides of the view back by the provided amount. This has been corrected by multiplying the "padding" by 2 for the right and bottom sides only.