Water running
Rail state now checks if the ground_inst exists to exit into running.
Added player_set_water_run which checks if the player is fast enough and flush with a body of water to keep running.
Restored a few comments in the Effects step block.
Renamed surface_time to surface_alarm.