Overhauled the player's ground alignment
Sonic Studio Engine's ground alignment algorithm is not very efficient. After the player has found a solid intersecting the upper/lower half of their collision mask, there are two more collision checks used to align them to that solid:
1. Check if the player's lower half is inside the solid, and if so, rise up from it.
2. Check if the player's lower half is outside the solid, and if so, snap down to it.

To make matters worse, all of these are performed in a loop, so it ultimately doesn't come out very fast.

Basic Sonic Engine and Sonic For GMS omit these collision checks in favor of directly using the height at which the solid has been found during the first collision check to align the player to the ground accordingly. The manner in which VectorSatyr achieved this was by moving the extension of the player's collision mask outside of the actual floor / ceiling collision functions, and integrating it into the relevant sections in the player's grounded and aerial movement step functions.

It is my desire to maintain compatibility with Sonic Studio Engine as much as possible, so I have taken a different approach to this without heavy modification:
1. A `ground_offset` variable has been initialized in the player's Create event, which is used to store the height at which a solid is found in their floor and ceiling collision functions.
2. The algorithm for aligning the player to their ground has been rewritten the same as Basic Sonic Engine and Sonic For GMS, utilizing this `ground_offset` variable.

As a result of these changes, the current method of rotating the player's collision mask (assigning their mask direction as their angle rounded to 90 degrees) is seemingly no longer compatible, especially with custom terrain shapes. Therefore, the `player_rotate_mask` function has been re-instated.

Due to the smaller collision mask size the player has, it was also necessary to increase the scanning height for angle calculation of custom terrain shapes (and by extension, the scanning height for recording the player's local instances.)

Various comments have also been rewritten for further clarity.