Re-added Sonic For GMS' terrain mapping fixes
The fixes for mapping to a given terrain within a certain number of degrees, as well as aligning to terrain based on the player's current rotation and not their mask direction, have been re-introduced following proper research into their functionality, as well as what they were intended to address; see: https://github.com/VectorSatyr/legacy-sonic-for-gms-1/issues/11

The out-of-bounds defaulting for the quarter-pipe shape specifically has been commented out: this is because it would cause the player to partially snap down to a quarter-pipe before falling off of it as they attempted to walk along it from a non-curved shape. However, this has a side effect: if the player falls exactly one pixel to the left of a quarter-pipe's top, they will, for lack of a better term, "slide down" it before falling properly. This is because, in Sonic For GMS' ground alignment algorithm, collision as valid at the top of the quarter-pipe where the player's angle is 90 degrees, so they are pushed forward horizontally. This is present in Sonic For GMS as well, but the player gets pushed forward instantly instead of gradually in this case (possibly because Sonic For GMS handles collision with terrain on a geometric-level instead of relying on mask intersections.)

The changes introduced in the "ground alignment overhaul" commit have had to be reverted in order for these changes to be possible.

Some additional validation checks have been made in regards to rotation of the player's mask, and running on water:
- If the player is unsuccessful in mapping to a floor, their `ground_inst` variable gets set to `noone`.
- The player's mask rotation function is now called in their grounded movement step only if their `ground_inst` is not `noone`.
- Inversely, running on water is now only possible if the player's `ground_inst` is `noone`.