Welcome back input menu
Put Keyboard and Gamepad into an Input menu. Again.
Renamed "input_gamepad_focus" to "input_background" as it's not like the other gamepad settings.

Added a few controller specific resets and setting reset for input background to be called from.

Setup menus so that they return the name of their input for the guide.